# docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Sentinel AML API",
        "description": "Transaction monitoring: ingestion, rule-based detection, alerts and case management. Roles: ANALYST (queue & disposition), SUPERVISOR (+ full PII, case closure, reopen), ADMIN (+ ingestion, rule/watchlist/FX configuration).",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "basic": []
        }
    ],
...
    "paths": {
...
        "/api/v1/ingest/accounts": {
            "post": {
                "tags": [
                    "Ingestion"
                ],
                "summary": "Bulk upsert accounts from a JSON array; customerRef must already exist",
                "operationId": "accounts",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/AccountRecord"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngestionResult"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/accounts/{id}/unfreeze": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Lift an automatic freeze (SUPERVISOR only, reason required, audited)",
                "description": "409 if the account is not frozen",
                "operationId": "unfreeze",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Unfreeze"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Accounts by status (e.g. status=FROZEN for accounts frozen after a sanctions hit)",
                "operationId": "list_5",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "FROZEN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "additionalProperties": {
                                            "type": "object"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "IngestionResult": {
                "type": "object",
                "properties": {
                    "batchId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "recordType": {
                        "type": "string"
                    },
                    "received": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "accepted": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "duplicates": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "rejected": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RowError"
                        }
                    },
                    "errorsTruncated": {
                        "type": "boolean"
                    },
                    "ingestMs": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "detection": {
                        "$ref": "#/components/schemas/Summary"
                    }
                }
            },
            "RowError": {
                "type": "object",
                "properties": {
                    "line": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "key": {
                        "type": "string"
                    },
                    "error": {
                        "type": "string"
                    }
                }
            },
            "Summary": {
                "type": "object",
                "properties": {
                    "evaluated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "skipped": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "transactionsWithHits": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "alertsCreated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "alertsUpdated": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "failures": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "elapsedMs": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "txnPerSecond": {
                        "type": "number",
                        "format": "double"
                    }
                }
            },
...
            "AccountRecord": {
                "required": [
                    "accountNumber",
                    "accountType",
                    "currency",
                    "customerRef",
                    "openedOn",
                    "riskRating"
                ],
                "type": "object",
                "properties": {
                    "accountNumber": {
                        "maxLength": 40,
                        "minLength": 0,
                        "type": "string"
                    },
                    "customerRef": {
                        "maxLength": 40,
                        "minLength": 0,
                        "type": "string"
                    },
                    "accountType": {
                        "pattern": "SAVINGS|CURRENT|BUSINESS|NRE|WALLET",
                        "type": "string"
                    },
                    "currency": {
                        "pattern": "[A-Z]{3}",
                        "type": "string"
                    },
                    "openedOn": {
                        "type": "string",
                        "format": "date"
                    },
                    "riskRating": {
                        "pattern": "LOW|MEDIUM|HIGH",
                        "type": "string"
                    },
                    "status": {
                        "pattern": "ACTIVE|DORMANT|FROZEN|CLOSED",
                        "type": "string"
                    }
                }
            },
...
            "Unfreeze": {
                "required": [
                    "reason"
                ],
                "type": "object",
                "properties": {
                    "reason": {
                        "maxLength": 1000,
                        "minLength": 10,
                        "type": "string"
                    }
                }
            },
            "Account": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "accountNumber": {
                        "type": "string"
                    },
                    "customerId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "accountType": {
                        "type": "string"
                    },
                    "currency": {
                        "type": "string"
                    },
                    "openedOn": {
                        "type": "string",
                        "format": "date"
                    },
                    "riskRating": {
                        "type": "string",
                        "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH"
                        ]
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "basic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}